Added token authorization in email group api endpoints

+++ email-group/controllers/delete.js
@@ -1,20 +1,67 @@
-const EmailGroup = require("../model");
+const mongoose = require('mongoose');
+
+const EmailGroup = require('./../model');
+const { tokenSchema } = require('../../model');
+
+const Token = new mongoose.model('Token', tokenSchema);
 
 const deleteEmailGroup = async (req, res) => {
-  try {
-    const emailGroupId = req.params.id;
+	try {
+		const authHeader = req.headers.authorization;
+		if (!authHeader.startsWith('Bearer ')) {
+			return res.status(400).json({
+				error: 'Invalid request headers.',
+			});
+		}
 
-    EmailGroup.deleteOne({ _id: emailGroupId }, (err) => {
-      if (err) {
-        return res.status(400).json({ error: err.message });
-      }
-      return res
-        .status(200)
-        .json({ success: "Email Group deleted successfully." });
-    });
-  } catch (error) {
-    return res.status(400).json({ error: error.message });
-  }
+		const tokenData = authHeader.split(' ')[1];
+		if (!tokenData) {
+			return res.status(400).json({
+				error: 'Invalid token.',
+			});
+		}
+
+		const token = await Token.findOne({ token: tokenData });
+		if (!token) {
+			return res.status(400).json({
+				error: 'Invalid token.',
+			});
+		}
+
+		const tokenOwner = token.userid;
+
+		const emailGroupId = req.params.id;
+		const emailGroup = await EmailGroup.findById(emailGroupId);
+
+		if (!emailGroup) {
+			return res.status(400).json({
+				error: 'Invalid request.',
+			});
+		}
+
+		console.log(emailGroup);
+
+		const ownerId = emailGroup.owner;
+		if (tokenOwner.toString() !== ownerId.toString()) {
+			console.log('Expected:', tokenOwner, typeof tokenOwner);
+			console.log('Found:', ownerId, typeof ownerId);
+
+			return res.status(400).json({
+				error: 'Unauthorized request.',
+			});
+		}
+
+		EmailGroup.deleteOne({ _id: emailGroupId }, (err) => {
+			if (err) {
+				return res.status(400).json({ error: err.message });
+			}
+			return res
+				.status(200)
+				.json({ success: 'Email Group deleted successfully.' });
+		});
+	} catch (error) {
+		return res.status(400).json({ error: error.message });
+	}
 };
 
 module.exports = deleteEmailGroup;
